move fetch to separate file
